Tweak animation clip editor behaviour by moving most logic to add new tracks to timeline
