+++ DTXMania/UI/Drawable/Animation/AnimationClipEditor.Inspector.cs
@@ -1,6 +1,4 @@
-using System.Numerics;
-using System.Reflection;
-using DTXMania.Core.Framework;
+using System.IO;
 using DTXMania.UI.Drawable;
 using Hexa.NET.ImGui;
 using NativeFileDialog.Extended;
@@ -29,9 +27,6 @@ public sealed partial class AnimationClipEditor
         {
             animator.Play(selectedClip.name);
         }
-
-        ImGui.Separator();
-        DrawTrackList(selectedClip, root);
 
         ImGui.Separator();
         if (ImGui.Button(timelineWindowOpen ? "Close Timeline" : "Open Timeline"))
@@ -174,207 +169,5 @@ public sealed partial class AnimationClipEditor
         ImGui.InputFloat("Duration", ref clip.duration);
         if (clip.duration < 0f) clip.duration = 0f;
         ImGui.Checkbox("Loop", ref clip.loop);
-    }
-
-    private void DrawTrackList(AnimationClip clip, UIGroup root)
-    {
-        ImGui.Text("Tracks");
-        ImGui.SameLine();
-
-        // The "+" track button opens a nested menu that walks the drawable tree and lists
-        // themable properties on each. Clicking a leaf adds a new track with the resolved path.
-        if (ImGui.BeginPopup("addtrackpopup"))
-        {
-            DrawAddTrackMenu(clip, root, parentPath: "");
-            ImGui.EndPopup();
-        }
-        if (ImGui.SmallButton("+##addtrack"))
-        {
-            ImGui.OpenPopup("addtrackpopup");
-        }
-
-        for (int i = 0; i < clip.tracks.Count; i++)
-        {
-            AnimationTrack track = clip.tracks[i];
-            ImGui.PushID(i);
-
-            bool isSelected = track == selectedTrack;
-            string label = string.IsNullOrEmpty(track.path) ? "(unbound)" : track.path;
-            if (ImGui.Selectable($"{label}##track", isSelected))
-            {
-                if (selectedTrack != track)
-                {
-                    selectedTrack = track;
-                    selectedKeyframe = null;
-                }
-            }
-            if (ImGui.BeginPopupContextItem("trackctx"))
-            {
-                if (ImGui.MenuItem("Delete"))
-                {
-                    clip.tracks.RemoveAt(i);
-                    if (selectedTrack == track) { selectedTrack = null; selectedKeyframe = null; }
-                    ImGui.EndPopup();
-                    ImGui.PopID();
-                    return;
-                }
-                ImGui.EndPopup();
-            }
-            ImGui.PopID();
-        }
-    }
-
-    // ---- Add Property menu ----------------------------------------------------------------
-
-    /// <summary>
-    /// Recursive menu builder. At each level we present, in order:
-    ///   1. Submenus for each child of the current drawable (recursively populated).
-    ///      Unnamed children are shown disabled, because a track path can't address them.
-    ///   2. A separator.
-    ///   3. Leaf items for each themable property on the current drawable.
-    /// </summary>
-    private void DrawAddTrackMenu(AnimationClip clip, UIDrawable current, string parentPath)
-    {
-        // 1) Children first.
-        bool hasAnyChildren = false;
-        if (current is UIGroup group && group.children.Count > 0)
-        {
-            for (int i = 0; i < group.children.Count; i++)
-            {
-                UIDrawable child = group.children[i];
-                hasAnyChildren = true;
-
-                if (string.IsNullOrEmpty(child.name))
-                {
-                    // Unnamed → not addressable. Show but disable so the user understands
-                    // they need to name it first.
-                    ImGui.BeginDisabled();
-                    ImGui.MenuItem($"(unnamed {child.GetType().Name})");
-                    ImGui.EndDisabled();
-                    continue;
-                }
-
-                string childLabel = $"{child.name} ({child.GetType().Name})";
-                ImGui.PushID(i);
-                if (ImGui.BeginMenu(childLabel))
-                {
-                    string childPath = string.IsNullOrEmpty(parentPath)
-                        ? child.name
-                        : parentPath + "/" + child.name;
-                    DrawAddTrackMenu(clip, child, childPath);
-                    ImGui.EndMenu();
-                }
-                ImGui.PopID();
-            }
-        }
-
-        // 2) Separator between children and own properties (only if both exist).
-        List<PropertyEntry> ownProperties = EnumerateThemableProperties(current.GetType()).ToList();
-        if (hasAnyChildren && ownProperties.Count > 0)
-        {
-            ImGui.Separator();
-        }
-
-        // 3) Own properties.
-        foreach (PropertyEntry entry in ownProperties)
-        {
-            string basePath = string.IsNullOrEmpty(parentPath)
-                ? entry.Name
-                : parentPath + "/" + entry.Name;
-
-            string[] subFields = GetAnimatableSubFields(entry.ValueType);
-            bool wholeAnimatable = Interpolator.IsRegistered(entry.ValueType);
-
-            if (subFields.Length == 0 && wholeAnimatable)
-            {
-                if (ImGui.MenuItem(entry.Name))
-                {
-                    AddTrack(clip, basePath);
-                    ImGui.CloseCurrentPopup();
-                }
-            }
-            else if (subFields.Length > 0)
-            {
-                if (ImGui.BeginMenu(entry.Name))
-                {
-                    if (wholeAnimatable && ImGui.MenuItem($"{entry.Name} (whole)"))
-                    {
-                        AddTrack(clip, basePath);
-                        ImGui.CloseCurrentPopup();
-                    }
-                    foreach (string sub in subFields)
-                    {
-                        if (ImGui.MenuItem(sub))
-                        {
-                            AddTrack(clip, basePath + "." + sub);
-                            ImGui.CloseCurrentPopup();
-                        }
-                    }
-                    ImGui.EndMenu();
-                }
-            }
-            else
-            {
-                // Themable but no known interpolator and no usable sub-fields. Greyed.
-                ImGui.BeginDisabled();
-                ImGui.MenuItem($"{entry.Name} ({entry.ValueType.Name})");
-                ImGui.EndDisabled();
-            }
-        }
-    }
-
-    private void AddTrack(AnimationClip clip, string path)
-    {
-        AnimationTrack track = new() { path = path };
-        clip.tracks.Add(track);
-        selectedTrack = track;
-        selectedKeyframe = null;
-    }
-
-    private readonly struct PropertyEntry
-    {
-        public PropertyEntry(string name, Type valueType) { Name = name; ValueType = valueType; }
-        public string Name { get; }
-        public Type ValueType { get; }
-    }
-
-    private static IEnumerable<PropertyEntry> EnumerateThemableProperties(Type type)
-    {
-        const BindingFlags flags = BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance;
-
-        // Fields
-        foreach (FieldInfo f in type.GetFields(flags))
-        {
-            if (HasThemable(f))
-                yield return new PropertyEntry(f.Name, f.FieldType);
-        }
-        // Properties (rare in your codebase but harmless to support)
-        foreach (PropertyInfo p in type.GetProperties(flags))
-        {
-            if (p.GetIndexParameters().Length > 0) continue;
-            if (HasThemable(p))
-                yield return new PropertyEntry(p.Name, p.PropertyType);
-        }
-    }
-
-    private static bool HasThemable(MemberInfo m)
-    {
-        foreach (var a in m.GetCustomAttributes(inherit: true))
-            if (a is ThemableAttribute) return true;
-        return false;
-    }
-
-    /// <summary>
-    /// For a property value type, return the names of sub-fields we want to offer as separate
-    /// animation targets. Vectors → X/Y/Z/W, Color4 → Red/Green/Blue/Alpha. Returns an empty
-    /// array for types where the whole value is the only sensible target.
-    /// </summary>
-    private static string[] GetAnimatableSubFields(Type valueType)
-    {
-        if (valueType == typeof(Vector2)) return ["X", "Y"];
-        if (valueType == typeof(Vector3)) return ["X", "Y", "Z"];
-        if (valueType == typeof(Vector4)) return ["X", "Y", "Z", "W"];
-        if (valueType == typeof(Color4)) return ["Red", "Green", "Blue", "Alpha"];
-        return [];
     }
 }
